Fix Bollinger Band overflow: cast AVG result to integer
Neon driver serializes PostgreSQL numeric type as JS string, causing
string concatenation instead of addition in the band calculator.
Adding ::integer cast ensures numeric type is returned.

+++ src/app/api/prices/route.ts
@@ -38,8 +38,8 @@ export async function POST() {
     }>(
       `SELECT
          MAX(timestamp) AS timestamp,
-         ROUND(AVG(gold_bar_sell)) AS gold_bar_sell,
-         ROUND(AVG(gold_bar_buy))  AS gold_bar_buy
+         ROUND(AVG(gold_bar_sell))::integer AS gold_bar_sell,
+         ROUND(AVG(gold_bar_buy))::integer  AS gold_bar_buy
        FROM price_history
        WHERE gold_bar_sell > 0
        GROUP BY DATE(timestamp AT TIME ZONE 'Asia/Bangkok')
